Organized tests and removed unnecessary ones
