+++ integration/glide.lock
@@ -1,0 +1,132 @@
+hash: 3571c3de47e367a5cd9e8fcc70ad6e342a2c6c6edb2212927f2112e858f04d0f
+updated: 2017-01-19T14:29:44.122057237-05:00
+imports:
+- name: github.com/aws/aws-sdk-go
+  version: 1355b456f6ba4a8453249e0dc7743f16ccac362d
+  subpackages:
+  - aws
+  - aws/credentials
+  - aws/session
+  - service/ec2
+  - service/route53
+  - aws/awserr
+  - service/efs
+  - aws/awsutil
+  - service/s3
+  - aws/endpoints
+  - aws/client
+  - aws/corehandlers
+  - aws/credentials/stscreds
+  - aws/defaults
+  - aws/request
+  - aws/client/metadata
+  - aws/signer/v4
+  - private/protocol
+  - private/protocol/ec2query
+  - private/waiter
+  - private/protocol/restjson
+  - private/protocol/restxml
+  - service/sts
+  - aws/credentials/ec2rolecreds
+  - aws/credentials/endpointcreds
+  - aws/ec2metadata
+  - private/protocol/rest
+  - private/protocol/query/queryutil
+  - private/protocol/xml/xmlutil
+  - private/protocol/jsonrpc
+  - private/protocol/query
+  - private/protocol/json/jsonutil
+- name: github.com/cloudflare/cfssl
+  version: db0d0650b6496bfe8061ec56a92edd32d8e75c30
+  subpackages:
+  - csr
+  - cli/genkey
+  - config
+  - helpers
+  - initca
+  - log
+  - signer
+  - signer/local
+  - errors
+  - cli
+  - auth
+  - ocsp/config
+  - crypto/pkcs7
+  - helpers/derhelpers
+  - certdb
+  - info
+  - signer/universal
+  - signer/remote
+  - api/client
+  - api
+- name: github.com/go-ini/ini
+  version: 766e555c68dc8bda90d197ee8946c37519c19409
+- name: github.com/google/certificate-transparency
+  version: 6e5648d13b43cdde25cc8274b87675230ffefbd6
+  repo: https://github.com/google/certificate-transparency
+  subpackages:
+  - go
+  - go/client
+  - go/x509
+  - go/asn1
+  - go/x509/pkix
+- name: github.com/jmespath/go-jmespath
+  version: bd40a432e4c76585ef6b72d3fd96fb9b6dc7b68d
+- name: github.com/mitchellh/go-homedir
+  version: b8bc1bf767474819792c23f32d8286a45736f1c6
+- name: github.com/mreiferson/go-httpclient
+  version: 63fe23f7434723dc904c901043af07931f293c47
+  repo: https://github.com/mreiferson/go-httpclient
+- name: github.com/onsi/ginkgo
+  version: a23f924ce96d61f963fb6219b8ff069d8d768cc2
+  subpackages:
+  - config
+  - internal/codelocation
+  - internal/failer
+  - internal/remote
+  - internal/suite
+  - internal/testingtproxy
+  - internal/writer
+  - reporters
+  - reporters/stenographer
+  - types
+  - internal/containernode
+  - internal/leafnodes
+  - internal/spec
+  - internal/specrunner
+  - reporters/stenographer/support/go-colorable
+  - reporters/stenographer/support/go-isatty
+- name: github.com/onsi/gomega
+  version: e85f63af4302dc0e4277c6008ecf803f1d80e4f0
+  subpackages:
+  - internal/assertion
+  - internal/asyncassertion
+  - internal/testingtsupport
+  - matchers
+  - types
+  - internal/oraclematcher
+  - format
+  - matchers/support/goraph/bipartitegraph
+  - matchers/support/goraph/edge
+  - matchers/support/goraph/node
+  - matchers/support/goraph/util
+- name: github.com/packethost/packngo
+  version: 91a3a4f65e3cd6ee66d919b4e7103ae038ce66ff
+- name: golang.org/x/crypto
+  version: 1f22c0103821b9390939b6776727195525381532
+  repo: https://go.googlesource.com/crypto
+  subpackages:
+  - ocsp
+  - pbkdf2
+  - pkcs12
+  - scrypt
+  - pkcs12/internal/rc2
+- name: golang.org/x/sys
+  version: d75a52659825e75fff6158388dddc6a5b04f9ba5
+  subpackages:
+  - unix
+- name: gopkg.in/yaml.v2
+  version: a5b47d31c556af34a302ce5d659e6fea44d90de0
+testImports:
+- name: gopkg.in/yaml.v2
+  version: a5b47d31c556af34a302ce5d659e6fea44d90de0
