+++ integration/install.go
@@ -2,17 +2,13 @@ package integration
 
 import (
 	"bufio"
-	"crypto/tls"
-	"fmt"
 	"html/template"
 	"log"
-	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"time"
 
-	"github.com/apprenda/kismatic/integration/retry"
 	homedir "github.com/mitchellh/go-homedir"
 	. "github.com/onsi/ginkgo"
 )
@@ -39,11 +35,6 @@ type installOptions struct {
 }
 
 func installKismaticMini(node NodeDeets, sshKey string) error {
-	By("Building a template")
-	template, err := template.New("planAWSOverlay").Parse(planAWSOverlay)
-	FailIfError(err, "Couldn't parse template")
-
-	By("Building a plan to set up an overlay network cluster on this hardware")
 	sshUser := node.SSHUser
 	plan := PlanAWS{
 		Etcd:                     []NodeDeets{node},
@@ -57,38 +48,11 @@ func installKismaticMini(node NodeDeets, sshKey string) error {
 		SSHUser:                  sshUser,
 		AllowPackageInstallation: true,
 	}
-
-	By("Writing plan file out to disk")
-	f, err := os.Create("kismatic-testing.yaml")
-	FailIfError(err, "Error waiting for nodes")
-	defer f.Close()
-	w := bufio.NewWriter(f)
-	err = template.Execute(w, &plan)
-	FailIfError(err, "Error filling in plan template")
-	w.Flush()
-
-	By("Validing our plan")
-	cmd := exec.Command("./kismatic", "install", "validate", "-f", f.Name())
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	err = cmd.Run()
-	FailIfError(err, "Error validating plan")
-
-	By("Punch it Chewie!")
-	cmd = exec.Command("./kismatic", "install", "apply", "-f", f.Name())
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	return cmd.Run()
+	return installKismaticWithPlan(plan, sshKey)
 }
 
 func installKismatic(nodes provisionedNodes, installOpts installOptions, sshKey string) error {
-	By("Building a template")
-	template, err := template.New("planAWSOverlay").Parse(planAWSOverlay)
-	FailIfError(err, "Couldn't parse template")
-
-	By("Building a plan to set up an overlay network cluster on this hardware")
 	sshUser := nodes.master[0].SSHUser
-
 	masterDNS := nodes.master[0].Hostname
 	if nodes.dnsRecord != nil && nodes.dnsRecord.Name != "" {
 		masterDNS = nodes.dnsRecord.Name
@@ -110,21 +74,7 @@ func installKismatic(nodes provisionedNodes, installOpts installOptions, sshKey 
 		DockerRegistryPort:           installOpts.dockerRegistryPort,
 		ModifyHostsFiles:             installOpts.modifyHostsFiles,
 	}
-
-	f, err := os.Create("kismatic-testing.yaml")
-	FailIfError(err, "Error creating plan")
-	defer f.Close()
-	w := bufio.NewWriter(f)
-	err = template.Execute(w, &plan)
-	FailIfError(err, "Error filling in plan template")
-	w.Flush()
-
-	By("Punch it Chewie!")
-	cmd := exec.Command("./kismatic", "install", "apply", "-f", f.Name())
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	return cmd.Run()
+	return installKismaticWithPlan(plan, sshKey)
 }
 
 func installKismaticWithPlan(plan PlanAWS, sshKey string) error {
@@ -146,89 +96,6 @@ func installKismaticWithPlan(plan PlanAWS, sshKey string) error {
 	cmd.Stderr = os.Stderr
 
 	return cmd.Run()
-}
-
-func verifyMasterNodeFailure(nodes provisionedNodes, provisioner infrastructureProvisioner, sshKey string) error {
-	By("Removing a Kubernetes master node")
-	if err := provisioner.TerminateNode(nodes.master[0]); err != nil {
-		return fmt.Errorf("Could not remove node: %v", err)
-	}
-
-	By("Rerunning Kuberang")
-	if err := runViaSSH([]string{"sudo kuberang"}, []NodeDeets{nodes.master[1]}, sshKey, 5*time.Minute); err != nil {
-		return fmt.Errorf("Failed to run kuberang: %v", err)
-	}
-
-	return nil
-}
-
-func verifyIngressNodes(master NodeDeets, ingressNodes []NodeDeets, sshKey string) error {
-	By("Adding a service and an ingress resource")
-	addIngressResource(master, sshKey)
-
-	By("Verifying the service is accessible via the ingress point(s)")
-	for _, ingNode := range ingressNodes {
-		if err := verifyIngressPoint(ingNode); err != nil {
-			// For debugging purposes...
-			runViaSSH([]string{"sudo kubectl describe -f /tmp/ingress.yaml", "sudo kubectl describe pods"}, []NodeDeets{master}, sshKey, 1*time.Minute)
-			return err
-		}
-	}
-
-	return nil
-}
-
-func addIngressResource(node NodeDeets, sshKey string) {
-	err := copyFileToRemote("test-resources/ingress.yaml", "/tmp/ingress.yaml", node, sshKey, 1*time.Minute)
-	FailIfError(err, "Error copying ingress test file")
-
-	err = runViaSSH([]string{"sudo openssl req -x509 -nodes -days 365 -newkey rsa:2048 -keyout /tmp/tls.key -out /tmp/tls.crt -subj \"/CN=kismaticintegration.com\""}, []NodeDeets{node}, sshKey, 1*time.Minute)
-	FailIfError(err, "Error creating certificates for HTTPs")
-
-	err = runViaSSH([]string{"sudo kubectl create secret tls kismaticintegration-tls --cert=/tmp/tls.crt --key=/tmp/tls.key"}, []NodeDeets{node}, sshKey, 1*time.Minute)
-	FailIfError(err, "Error creating tls secret")
-
-	err = runViaSSH([]string{"sudo kubectl apply -f /tmp/ingress.yaml"}, []NodeDeets{node}, sshKey, 1*time.Minute)
-	FailIfError(err, "Error creating ingress resources")
-}
-
-func verifyIngressPoint(node NodeDeets) error {
-	// HTTP ingress
-	url := "http://" + node.PublicIP + "/echo"
-	if err := retry.WithBackoff(func() error { return ingressRequest(url) }, 7); err != nil {
-		return err
-	}
-	// HTTPS ingress
-	url = "https://" + node.PublicIP + "/echo-tls"
-	if err := retry.WithBackoff(func() error { return ingressRequest(url) }, 7); err != nil {
-		return err
-	}
-	return nil
-}
-
-func ingressRequest(url string) error {
-	tr := &http.Transport{
-		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-	}
-	client := http.Client{
-		Timeout:   1000 * time.Millisecond,
-		Transport: tr,
-	}
-	req, err := http.NewRequest(http.MethodGet, url, nil)
-	if err != nil {
-		return fmt.Errorf("Could not create request for ingress via %s, %v", url, err)
-	}
-	// Set the host header since this is not a real domain, curl $IP/echo -H 'Host: kismaticintegration.com'
-	req.Host = "kismaticintegration.com"
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("Could not reach ingress via %s, %v", url, err)
-	}
-	if resp.StatusCode != 200 {
-		return fmt.Errorf("Ingress status code is not 200, got %d vi %s", resp.StatusCode, url)
-	}
-
-	return nil
 }
 
 func installKismaticWithABadNode() {
