+++ integration/install_released_test.go
@@ -16,34 +16,34 @@ var _ = Describe("Installing with previous version of Kismatic", func() {
 		// setup previous version of Kismatic
 		var err error
 		kisReleasedPath, err = DownloadKismaticRelease(previousKismaticVersion)
-		if err != nil {
-			Fail("Failed to download kismatic released")
-		}
+		Expect(err).ToNot(HaveOccurred(), "Failed to download kismatic release")
 		os.Chdir(kisReleasedPath)
 	})
+
 	AfterEach(func() {
 		if !leaveIt() {
 			os.RemoveAll(kisReleasedPath)
 		}
 	})
+
 	installOpts := installOptions{
 		allowPackageInstallation: true,
 	}
-	Context("Targeting AWS infrastructure", func() {
-		Context("using a 1/1/1 layout with Ubuntu 16.04 LTS", func() {
-			ItOnAWS("should result in a working cluster", func(provisioner infrastructureProvisioner) {
-				WithInfrastructure(NodeCount{1, 1, 1, 0, 0}, Ubuntu1604LTS, provisioner, func(nodes provisionedNodes, sshKey string) {
-					err := installKismatic(nodes, installOpts, sshKey)
-					Expect(err).ToNot(HaveOccurred())
-				})
+
+	Context("using Ubuntu 16.04 LTS", func() {
+		ItOnAWS("should install successfully", func(aws infrastructureProvisioner) {
+			WithInfrastructure(NodeCount{1, 1, 1, 0, 0}, Ubuntu1604LTS, aws, func(nodes provisionedNodes, sshKey string) {
+				err := installKismatic(nodes, installOpts, sshKey)
+				Expect(err).ToNot(HaveOccurred())
 			})
 		})
-		Context("using a 1/1/1 layout with CentOS 7", func() {
-			ItOnAWS("should result in a working cluster", func(provisioner infrastructureProvisioner) {
-				WithInfrastructure(NodeCount{1, 1, 1, 0, 0}, Ubuntu1604LTS, provisioner, func(nodes provisionedNodes, sshKey string) {
-					err := installKismatic(nodes, installOpts, sshKey)
-					Expect(err).ToNot(HaveOccurred())
-				})
+	})
+
+	Context("using CentOS", func() {
+		ItOnAWS("should install successfully", func(aws infrastructureProvisioner) {
+			WithInfrastructure(NodeCount{1, 1, 1, 0, 0}, CentOS7, aws, func(nodes provisionedNodes, sshKey string) {
+				err := installKismatic(nodes, installOpts, sshKey)
+				Expect(err).ToNot(HaveOccurred())
 			})
 		})
 	})
