+++ integration/kismatic_platform_suite_test.go
@@ -9,6 +9,8 @@ import (
 	"path/filepath"
 	"time"
 
+	"testing"
+
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awsutil"
 	"github.com/aws/aws-sdk-go/aws/credentials"
@@ -16,8 +18,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	"testing"
 )
 
 func TestKismaticPlatform(t *testing.T) {
