+++ integration/step_test.go
@@ -12,9 +12,10 @@ var _ = Describe("install step commands", func() {
 	BeforeEach(func() {
 		os.Chdir(kisPath)
 	})
+
 	Context("Running the api server play against an existing cluster", func() {
-		ItOnAWS("should return successfully", func(provisioner infrastructureProvisioner) {
-			WithMiniInfrastructure(CentOS7, provisioner, func(node NodeDeets, sshKey string) {
+		ItOnAWS("should return successfully", func(aws infrastructureProvisioner) {
+			WithMiniInfrastructure(CentOS7, aws, func(node NodeDeets, sshKey string) {
 				err := installKismaticMini(node, sshKey)
 				Expect(err).ToNot(HaveOccurred())
 
@@ -25,9 +26,10 @@ var _ = Describe("install step commands", func() {
 				Expect(err).ToNot(HaveOccurred())
 			})
 		})
+
 		Context("Running the smoketest play against an existing cluster", func() {
-			ItOnAWS("should return successfully", func(provisioner infrastructureProvisioner) {
-				WithMiniInfrastructure(CentOS7, provisioner, func(node NodeDeets, sshKey string) {
+			ItOnAWS("should return successfully", func(aws infrastructureProvisioner) {
+				WithMiniInfrastructure(CentOS7, aws, func(node NodeDeets, sshKey string) {
 					err := installKismaticMini(node, sshKey)
 					Expect(err).ToNot(HaveOccurred())
