+++ integration/validate.go
@@ -2,13 +2,56 @@ package integration
 
 import (
 	"bufio"
+	"fmt"
 	"html/template"
 	"log"
 	"os"
 	"os/exec"
+	"path/filepath"
+	"time"
 
+	homedir "github.com/mitchellh/go-homedir"
 	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
 )
+
+func validateMiniPkgInstallationDisabled(provisioner infrastructureProvisioner, distro linuxDistro) {
+	WithMiniInfrastructure(distro, provisioner, func(node NodeDeets, sshKey string) {
+		sshUser := node.SSHUser
+		if err := ValidateKismaticMiniDenyPkgInstallation(node, sshUser, sshKey); err == nil {
+			Fail("Missing dependencies, but still passed")
+		}
+
+		By("Prepping nodes for the test")
+		prep := getPrepForDistro(distro)
+		prepNode := []NodeDeets{node}
+		err := runViaSSH(prep.CommandsToPrepRepo, prepNode, sshKey, 5*time.Minute)
+		FailIfError(err, "Failed to prep repo on the node")
+
+		By("Installing etcd on the node")
+		err = runViaSSH(prep.CommandsToInstallEtcd, prepNode, sshKey, 10*time.Minute)
+		FailIfError(err, "Failed to install etcd on the node")
+
+		if err = ValidateKismaticMiniDenyPkgInstallation(node, sshUser, sshKey); err == nil {
+			Fail("Missing dependencies, but still passed")
+		}
+
+		By("Installing Docker")
+		err = runViaSSH(prep.CommandsToInstallDocker, prepNode, sshKey, 10*time.Minute)
+		FailIfError(err, "failed to install docker over SSH")
+
+		if err = ValidateKismaticMiniDenyPkgInstallation(node, sshUser, sshKey); err == nil {
+			Fail("Missing dependencies, but still passed")
+		}
+
+		By("Installing Master")
+		err = runViaSSH(prep.CommandsToInstallK8sMaster, prepNode, sshKey, 15*time.Minute)
+		FailIfError(err, "Failed to install master on node via SSH")
+
+		err = ValidateKismaticMiniDenyPkgInstallation(node, sshUser, sshKey)
+		Expect(err).To(BeNil())
+	})
+}
 
 // ValidateKismaticMini runs validation against a mini Kubernetes cluster
 func ValidateKismaticMini(node NodeDeets, user, sshKey string) PlanAWS {
@@ -118,4 +161,18 @@ func ValidateKismaticMiniWithBadSSH(node NodeDeets, user, sshKey string) PlanAWS
 	err = ver.Run()
 	FailIfSuccess(err)
 	return plan
+}
+
+func getBadSSHKeyFile() (string, error) {
+	dir, err := homedir.Dir()
+	if err != nil {
+		return "", err
+	}
+	// create empty file
+	_, err = os.Create(filepath.Join(dir, ".ssh", "bad.pem"))
+	if err != nil {
+		return "", fmt.Errorf("Unable to create tag file!")
+	}
+
+	return filepath.Join(dir, ".ssh", "bad.pem"), nil
 }
